requirements.txt
bs4
requests
schedule
pyTelegramBotAPI
openai
python-dotenv
PyMuPDF

